requirements.txt
flask
pandas
docx2txt
spacy
fuzzywuzzy
pdfminer.six
